// SDKs/twilio.com/twilio_events_v1/1.37.3/typescript/src/sdk/models/shared/eventsv1eventtype.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class EventsV1EventType extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=description" })
  description?: string;

  @SpeakeasyMetadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @SpeakeasyMetadata({ data: "json, name=schema_id" })
  schemaId?: string;

  @SpeakeasyMetadata({ data: "json, name=type" })
  type?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}